iroiro

+++ app/Http/Controllers/TasksController.php
@@ -15,10 +15,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index',[
-            'tasks' => $tasks,    
-        ]);
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+                
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index',[
+                'user' => $user,
+                'tasks' => $tasks,
+            ]);
+        }else{
+
+        // Welcomeビューでそれらを表示
+        return view('welcome');
+        }
     }
 
     /**
@@ -48,11 +63,10 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -115,7 +129,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-        return redirect('/');
+        
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
+    
+    return redirect('/');
     }
 }
